refactor(cdkactions): rename subpath imports for Bun compatibility
Bun cannot resolve subpath imports with `$` in the specifier. Rename
`#@/*` → `#src/*` and `#$/*` → `#test/*.js` (with explicit .js → .ts
remapping) so that `bun test` resolves all modules correctly.
