+++ packages/cdkactions/examples/01-nodejs-ci-matrix.ts
@@ -1,4 +1,4 @@
-import { App, Stack, Workflow, Job, RunnerLabel, expression } from '#@/index.js';
+import { App, Stack, Workflow, Job, RunnerLabel, expression } from '#src/index.js';
 import { checkoutV4, setupNodeV6 } from '../src/actions.js';
 
 const { matrix } = expression;
