+++ packages/cdkactions/examples/08-composite-action.ts
@@ -1,4 +1,4 @@
-import { App, Stack, Workflow, Job, RunnerLabel, Shell, CompositeAction, hashFiles } from '#@/index.js';
+import { App, Stack, Workflow, Job, RunnerLabel, Shell, CompositeAction, hashFiles } from '#src/index.js';
 
 export function create(app?: App) {
   const _app = app ?? new App();
