+++ packages/cdkactions/examples/10-cross-workflow-deps.ts
@@ -1,4 +1,4 @@
-import { App, Stack, Workflow, Job, RunnerLabel } from '#@/index.js';
+import { App, Stack, Workflow, Job, RunnerLabel } from '#src/index.js';
 
 export function create(app?: App) {
   const _app = app ?? new App();
