+++ packages/cdkactions/examples/15-runner-registry.ts
@@ -1,4 +1,4 @@
-import { App, Stack, Workflow, Job, RunnerLabel } from '#@/index.js';
+import { App, Stack, Workflow, Job, RunnerLabel } from '#src/index.js';
 
 const Runners = {
   ...RunnerLabel,
