+++ packages/cdkactions/src/actions.ts
@@ -1,4 +1,4 @@
-import { defineAction } from '#@/action.js';
+import { defineAction } from '#src/action.js';
 
 export const checkoutV2 = defineAction<{
   repository: { default: string };
