+++ packages/cdkactions/src/index.ts
@@ -1,12 +1,12 @@
-export * from '#@/action.js';
-export * from '#@/actions.js';
-export * from '#@/app.js';
-export * from '#@/composite-action.js';
-export * from '#@/expressions.js';
-export * from '#@/job.js';
-export * from '#@/library.js';
-export * from '#@/nominal.js';
-export * from '#@/stack.js';
-export * from '#@/types.js';
-export * from '#@/validation.js';
-export * from '#@/workflow.js';
+export * from '#src/action.js';
+export * from '#src/actions.js';
+export * from '#src/app.js';
+export * from '#src/composite-action.js';
+export * from '#src/expressions.js';
+export * from '#src/job.js';
+export * from '#src/library.js';
+export * from '#src/nominal.js';
+export * from '#src/stack.js';
+export * from '#src/types.js';
+export * from '#src/validation.js';
+export * from '#src/workflow.js';
