+++ packages/cdkactions/src/library.ts
@@ -1,12 +1,12 @@
 import type { Construct } from 'constructs';
 import { dedent } from 'ts-dedent';
 
-import { checkoutV4 } from '#@/actions.js';
-import { always, expr, github, secrets } from '#@/expressions.js';
-import { Job, type JobProps, type MatrixDefinition, type StepConfig } from '#@/job.js';
-import { RunnerLabel } from '#@/nominal.js';
-import { Stack } from '#@/stack.js';
-import { Workflow } from '#@/workflow.js';
+import { checkoutV4 } from '#src/actions.js';
+import { always, expr, github, secrets } from '#src/expressions.js';
+import { Job, type JobProps, type MatrixDefinition, type StepConfig } from '#src/job.js';
+import { RunnerLabel } from '#src/nominal.js';
+import { Stack } from '#src/stack.js';
+import { Workflow } from '#src/workflow.js';
 
 /**
  * Configuration for a CDKActionsStack instance.
