+++ packages/cdkactions/src/utils.ts
@@ -1,4 +1,4 @@
-import type { StringMap } from '#@/types.js';
+import type { StringMap } from '#src/types.js';
 
 export type Writable<T> = T extends Function ? T : T extends object ? { -readonly [K in keyof T]: Writable<T[K]> } : T;
